pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.13.1",
    "asyncpg>=0.30.0",
    "numpy>=2.3.4",
    "pandas>=2.3.3",
    "psycopg2-binary>=2.9.11",
    "python-binance>=1.0.30",
    "python-dotenv>=1.2.1",
    "python-telegram-bot>=22.5",
    "pytz>=2025.2",
    "redis>=7.0.1",
    "sqlalchemy>=2.0.44",
    "websockets>=15.0.1",
]
